Improving temperature sensor GUI
Added the function that enables the user to selest the serial port, baud rate before the start of measurement. Important changes in the __init__() becasue the serial read function starts only when the user establisehes the serial connection. Sometimes a bug appears that makes the plot rugged, step-like, not sure if it's fixed or not because it seemingly appears randomly.